resnet/requirements.txt
pillow==12.1.1
torch==2.10.0
torchvision==0.25.0
